# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/login": {
      "get": {
        "operationId": "AppController_getLogin",
        "parameters": [
          {
            "name": "error",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login page rendered"
          },
          "302": {
            "description": "Redirected if already logged in"
          }
        },
        "summary": "Render login page",
        "tags": [
          "App"
        ]
      }
    },
    "/editor": {
      "get": {
        "operationId": "AppController_getEditor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Editor rendered"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient role"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Render protected editor page",
        "tags": [
          "App"
        ]
      }
    },
    "/logout": {
      "get": {
        "operationId": "AppController_getLogout",
        "parameters": [
          {
            "name": "error",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logout page rendered"
          },
          "302": {
            "description": "Redirected if already logged in"
          }
        },
        "summary": "Render logout page",
        "tags": [
          "App"
        ]
      },
      "post": {
        "operationId": "AppController_handleLogout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User logged out successfully and redirected"
          }
        },
        "summary": "Log out a user and clear JWT cookie",
        "tags": [
          "App"
        ]
      }
    },
...
    "/api/gemini/file/generate-text": {
      "post": {
        "operationId": "GoogleGeminiFileController_generateText",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateTextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated text content.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Generate text from a simple text prompt.",
        "tags": [
          "Gemini AI"
        ]
      }
    },
    "/api/gemini/file/generate-image-base64": {
      "post": {
        "operationId": "GoogleGeminiFileController_generateTextWithBase64Image",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateImageBase64Dto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Generated text content based on the image and text prompt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Generate text from a prompt with an embedded Base64 image.",
        "tags": [
          "Gemini AI"
        ]
      }
    },
    "/api/gemini/file/generate-file": {
      "post": {
        "operationId": "GoogleGeminiFileController_generateTextWithFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Prompt text and the file to be analyzed.",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "The text prompt for Gemini.",
                    "example": "Analyze this SQL schema and suggest improvements."
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The file to attach (e.g., .sql, .txt, .csv)."
                  }
                },
                "required": [
                  "prompt",
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Generated text content based on the file and text prompt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Generate text from a prompt with an uploaded file (e.g., .sql, .txt).",
        "tags": [
          "Gemini AI"
        ]
      }
    },
...
    "/feature/status/{moduleName}": {
      "get": {
        "operationId": "FeatureController_getModuleStatus",
        "parameters": [
          {
            "name": "moduleName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Feature"
        ]
      }
    },
    "/feature/enable/{moduleName}": {
      "get": {
        "operationId": "FeatureController_enableModule",
        "parameters": [
          {
            "name": "moduleName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Feature"
        ]
      }
    },
    "/feature/disable/{moduleName}": {
      "get": {
        "operationId": "FeatureController_disableModule",
        "parameters": [
          {
            "name": "moduleName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Feature"
        ]
      }
    },
    "/feature/toggle/{moduleName}": {
      "get": {
        "operationId": "FeatureController_toggleModule",
        "parameters": [
          {
            "name": "moduleName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Feature"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Server",
    "description": "Authentication and Role Protected APIs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in the format: Bearer <token>",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT stored in cookie (for web clients)"
      }
    },
    "schemas": {
...
      "GenerateTextDto": {
        "type": "object",
        "properties": {
          "prompt": {
            "type": "string",
            "description": "The prompt text to send to the Gemini model.",
            "example": "Write a short story about a brave knight."
          }
        },
        "required": [
          "prompt"
        ]
      },
      "GenerateImageBase64Dto": {
        "type": "object",
        "properties": {
          "prompt": {
            "type": "string",
            "description": "The prompt text to send to the Gemini model, accompanying the image.",
            "example": "Describe this image."
          },
          "base64Image": {
            "type": "string",
            "description": "Base64 encoded image data (e.g., \"data:image/jpeg;base64,...\").",
            "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEAAAAAAAD/2wBDAAY..."
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type of the image (e.g., \"image/jpeg\", \"image/png\").",
            "example": "image/jpeg"
          }
        },
        "required": [
          "prompt",
          "base64Image",
          "mimeType"
        ]
      },
...
    }
  }
}